requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
tqdm>=4.60.0
numpy>=1.21.0
pandas>=1.3.0
faiss-gpu>=1.7.0
orjson>=3.6.0